requirements.txt
streamlit
jiwer
python-Levenshtein